fix(ui): guard negative idx casts + widen ID parse to i64
- history_model.rs: parse entry ID as i64 before narrowing to i32,
  preventing parse failure on IDs > i32::MAX from the daemon.
- root_window.rs: add `if idx < 0 { return; }` guard before every
  `idx as usize` cast in on_item_clicked, on_item_copy, on_item_pin,
  and on_item_delete callbacks — avoids wrapping to huge usize and
  an out-of-bounds panic when Slint passes the default -1 sentinel.

+++ crates/copypaste-ui/src/history_model.rs
@@ -41,7 +41,7 @@ fn content_type_to_kind(content_type: &str) -> slint::SharedString {
 /// Convert a [`crate::ipc_client::HistoryEntry`] to a Slint [`ClipItem`].
 fn entry_to_clip_item(entry: &crate::ipc_client::HistoryEntry) -> ClipItem {
     ClipItem {
-        id: entry.id.parse::<i32>().unwrap_or(0),
+        id: entry.id.parse::<i64>().unwrap_or(0) as i32,
         preview: entry.preview.clone().into(),
         kind: content_type_to_kind(&entry.content_type),
         wall_time: format_wall_time(entry.wall_time).into(),
